requirements.txt
pdf2docx>=0.5.6
docx2pdf>=0.1.8       # только для Windows
Pillow>=10.0.0
colorama>=0.4.6       # для цветного вывода в консоль (по желанию)
tqdm>=4.66.0          # можно использовать для прогресс-бара (по желанию)
